Begining update to 7.8+
Went through diff between current development book and my
most recent scripts and modified to fit any changes. I'm going
to add a "source ../lfs_profile" line to the beginning of every
script, and replace all program versions with a variable
VERSION-<program> set in lfs_profile. The versions are already
set, I just need to replace them in the scripts. This way any
version updates will be handled automatically when I update
them in lfs_profile (there are some scattered references of
packages in other packages that can be easy to miss, not to
mention the convenience of only having to change these in
one place!)